Add agent config parameters back & support systemd
Both the 1.5 deb and rpm packages try to use xinetd if it's installed or
systemd sockets otherwise if available.  This module will override this
behaviour and default to using systemd if supported by the OS.  To force
the use of xinetd, set `use_xinetd => true`.

When using systemd, drop-in files are used to override the configuration
provided by the the packages.