Add `encryption_secret` acceptance tests
* Make file content `Sensitive`
* Fix issue on Centos 6 by unsetting environment var LC_ALL in xinetd
config.  Error was `/bin/bash: warning: setlocale: LC_ALL: cannot change locale (en_US): No such file or directory`
This was forming part of the data and then couldn't be decrypted in the
tests by openssl.  Unsetting the var in xinetd is harmless as
check-mk-agent (which is a bash script), sets it to `C` in the first few lines.